Correctly detect VT support on Windows
