--- pkg/iostreams/color_unix.go
@@ -1,24 +1,0 @@
-//+build !windows
-
-package iostreams
-
-import (
-	"os"
-	"strings"
-)
-
-func Is256ColorSupported() bool {
-	return IsTrueColorSupported() ||
-		strings.Contains(os.Getenv("TERM"), "256") ||
-		strings.Contains(os.Getenv("COLORTERM"), "256")
-}
-
-func IsTrueColorSupported() bool {
-	term := os.Getenv("TERM")
-	colorterm := os.Getenv("COLORTERM")
-
-	return strings.Contains(term, "24bit") ||
-		strings.Contains(term, "truecolor") ||
-		strings.Contains(colorterm, "24bit") ||
-		strings.Contains(colorterm, "truecolor")
-}
