+++ pkg/iostreams/console.go
@@ -2,4 +2,12 @@
 
 package iostreams
 
-func (s *IOStreams) EnableVirtualTerminalProcessing() {}
+import "os"
+
+func (s *IOStreams) EnableVirtualTerminalProcessing() error {
+	return nil
+}
+
+func enableVirtualTerminalProcessing(f *os.File) error {
+	return nil
+}
