--- utils/table_printer_unix.go
@@ -1,7 +1,0 @@
-//+build !windows
-
-package utils
-
-func (t *ttyTablePrinter) availableWidth() int {
-	return t.maxWidth
-}
